+++ Controls/paging.ts
@@ -15,4 +15,4 @@
  */
 
 export {default as Paging} from './_paging/Paging';
-export {IArrowState} from './_paging/Paging';
+export {IArrowState} from './_paging/interface/IArrowState';
